+++ src/main/java/com/example/hospitalreservation/reservation/presentation/rest/ReservationRestController.java
@@ -1,41 +1,26 @@
-package com.example.hospitalreservation.reservation.presentation;
+package com.example.hospitalreservation.reservation.presentation.rest;
 
 import com.example.hospitalreservation.common.exception.ApplicationException;
-import com.example.hospitalreservation.reservation.presentation.request.CreateReservationRequest;
-import com.example.hospitalreservation.reservation.presentation.request.DeleteReservationRequest;
-import com.example.hospitalreservation.reservation.domain.entity.Reservation;
-import com.example.hospitalreservation.reservation.exception.ReservationExceptionCode;
 import com.example.hospitalreservation.reservation.application.ReservationService;
 import com.example.hospitalreservation.reservation.application.command.CreateReservationCommand;
 import com.example.hospitalreservation.reservation.application.command.DeleteReservationCommand;
+import com.example.hospitalreservation.reservation.domain.repository.ReservationRepository;
+import com.example.hospitalreservation.reservation.exception.ReservationExceptionCode;
+import com.example.hospitalreservation.reservation.presentation.dto.request.CreateReservationRequest;
+import com.example.hospitalreservation.reservation.presentation.dto.request.DeleteReservationRequest;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @Slf4j
 @RequestMapping("/reservations")
-@Controller
-public class ReservationController {
+@RestController
+public class ReservationRestController {
 
     private final ReservationService reservationService;
 
-    public ReservationController(ReservationService reservationService) {
+    public ReservationRestController(ReservationService reservationService) {
         this.reservationService = reservationService;
-    }
-
-    @GetMapping
-    public String getReservations(Model model) {
-        List<Reservation> reservations = reservationService.getAllReservations();
-        model.addAttribute("reservations", reservations);
-        return "index";
-    }
-
-    @GetMapping("/new")
-    public String showReservationForm() {
-        return "reservation_form";
     }
 
     @PostMapping
